API redesign
The API now consists of the methods `new` and `next`.

The local checkpoint and SPK inventory are provided to
the constructor. `next` is now responsible for locating
the PoA.

The next filter to fetch is determined by the `next_block_hash`
field of `GetBlockHeaderResult`. If the next header has negative
confirmations due to a reorg, we rewind the internal state until
we find a header still in the best chain.

Matching blocks contain the most up to date checkpoint, so it
can be applied immediately to the local chain.

Added `Tip` variant to `Event` enum for emitting the tip
checkpoint when all blocks have been scanned.

Removed `EventInner`.